+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Add `SpecPredicate` and `tag_maybe` to the public interface (#49)
 
 ## [v0.2.4] - 2019-12-19
 ### Added